pimd: [fix] send pim prune via correct interface when rpf upstream interface for a source changes
When the rpf upstream interface for a joined source changes, a prune was being sent
out the new rpf upstream interface that was intended for the old upstream router.
The prune should be sent out the old rpf upstream interface so that the old route
for the multicast source is pruned.

Reviewed-by: Jafar Al-Gharaibeh <jafar@atcorp.com>
Signed-off-by: Taylor Bouvin <tbouvin@atcorp.com>